Chatbots/Simple_Chatbot/requirements.txt
streamlit
python-dotenv
openai==0.28
